feat(v-select): refactored select process
completedly refactored and abstracted the select functionality to resolve a number of usability and
process issues
