+++ src/stylus/settings/_variables.styl
@@ -203,8 +203,6 @@ $chip-outline-color := $grey.base
 $chip-height := 32px
 $chip-margin := 4px
 $chip-small-height := 26px
-$chip-darken-background := rgba(0, 0, 0, 0.24)
-$chip-darken-background-outline := rgba(0, 0, 0, 0.06)
 
 // Collapsible
 $expansion-panel-collapse-height := 48px
